Move openFile to the new class
