+++ src/Controllers/Export/ExportController.php
@@ -249,7 +249,7 @@ final readonly class ExportController implements InvocableController
 
         // Open file on server if needed
         if ($saveOnServer) {
-            $message = $this->export->openFile(
+            $message = $this->export->outputHandler->openFile(
                 $filename,
                 $isQuickExport,
                 $request->getParsedBodyParam('quick_export_onserver_overwrite') === 'saveitover',
